+++ app.js
@@ -2,6 +2,9 @@ const dotenv = require("dotenv").config();
 
 const localDB = process.env.LOCAL_JSONFILE
 
-const uiCli = require('./code/input');
+const uiCli = require('./src/input');
+
+
+console.log(localDB)
